requirements.txt
requests>=2.28.0
pyyaml>=6.0
click>=8.0
rich>=12.0
tabulate>=0.9.0